requirements.txt
numpy
torch
pillow
pysimplegui
click
requests
pyspng
ninja
imageio-ffmpeg
dlib